move huge boilerplate to conftest
