+++ src/artemis/experiment_plans/tests/conftest.py
@@ -1,13 +1,24 @@
-from unittest.mock import MagicMock, patch
+from unittest.mock import DEFAULT, MagicMock, patch
 
 import pytest
 from bluesky.run_engine import RunEngine
+from bluesky.utils import Msg
 from dodal.beamlines import i03
 from dodal.devices.aperturescatterguard import AperturePositions
+from dodal.devices.attenuator import Attenuator
+from dodal.devices.backlight import Backlight
+from dodal.devices.detector_motion import DetectorMotion
+from dodal.devices.eiger import EigerDetector
+from dodal.devices.flux import Flux
+from dodal.devices.s4_slit_gaps import S4SlitGaps
 from dodal.devices.smargon import Smargon
+from dodal.devices.synchrotron import Synchrotron
+from dodal.devices.undulator import Undulator
+from dodal.devices.zebra import Zebra
 from ophyd.status import Status
 
 from artemis.experiment_plans.fast_grid_scan_plan import FGSComposite
+from artemis.experiment_plans.rotation_scan_plan import rotation_scan_plan
 from artemis.external_interaction.callbacks.fgs.fgs_callback_collection import (
     FGSCallbackCollection,
 )
@@ -137,6 +148,33 @@ def test_full_grid_scan_params():
     return GridScanWithEdgeDetectInternalParameters(**params)
 
 
+@pytest.fixture()
+def fake_create_devices(
+    eiger: EigerDetector,
+    smargon: Smargon,
+    zebra: Zebra,
+):
+    eiger.stage = MagicMock()
+    eiger.unstage = MagicMock()
+    mock_omega_sets = MagicMock(return_value=Status(done=True, success=True))
+
+    mock_arm_disarm = MagicMock(
+        side_effect=zebra.pc.arm.armed.set, return_value=Status(done=True, success=True)
+    )
+    zebra.pc.arm.set = mock_arm_disarm
+    smargon.omega.velocity.set = mock_omega_sets
+    smargon.omega.set = mock_omega_sets
+
+    devices = {
+        "eiger": i03.eiger(fake_with_ophyd_sim=True),
+        "smargon": smargon,
+        "zebra": zebra,
+        "detector_motion": i03.detector_motion(fake_with_ophyd_sim=True),
+        "backlight": i03.backlight(fake_with_ophyd_sim=True),
+    }
+    return devices
+
+
 @pytest.fixture
 def fake_fgs_composite(smargon: Smargon, test_fgs_params: InternalParameters):
     fake_composite = FGSComposite(
@@ -191,3 +229,196 @@ def mock_rotation_subscriptions(test_rotation_params):
     ):
         subscriptions = RotationCallbackCollection.from_params(test_rotation_params)
     return subscriptions
+
+
+def fake_read(obj, initial_positions, _):
+    initial_positions[obj] = 0
+    yield Msg("null", obj)
+
+
+def do_rotation_plan_for_tests(
+    run_engine,
+    callbacks,
+    sim_und,
+    sim_synch,
+    sim_slits,
+    sim_flux,
+    sim_att,
+    expt_params,
+    sim_sgon,
+    sim_zeb,
+    sim_bl,
+    sim_det,
+):
+    with (
+        patch(
+            "bluesky.preprocessors.__read_and_stash_a_motor",
+            fake_read,
+        ),
+        patch(
+            "artemis.experiment_plans.rotation_scan_plan.RotationCallbackCollection.from_params",
+            lambda _: callbacks,
+        ),
+        patch("dodal.beamlines.i03.undulator", lambda: sim_und),
+        patch("dodal.beamlines.i03.synchrotron", lambda: sim_synch),
+        patch("dodal.beamlines.i03.s4_slit_gaps", lambda: sim_slits),
+        patch("dodal.beamlines.i03.flux", lambda: sim_flux),
+        patch("dodal.beamlines.i03.attenuator", lambda: sim_att),
+    ):
+        run_engine(
+            rotation_scan_plan(
+                expt_params,
+                sim_sgon,
+                sim_zeb,
+                sim_bl,
+                sim_att,
+                sim_det,
+            )
+        )
+
+
+def setup_and_run_rotation_plan_for_tests(
+    RE: RunEngine,
+    test_params: RotationInternalParameters,
+    smargon: Smargon,
+    zebra: Zebra,
+    eiger: EigerDetector,
+    attenuator: Attenuator,
+    detector_motion: DetectorMotion,
+    backlight: Backlight,
+    mock_rotation_subscriptions: RotationCallbackCollection,
+    synchrotron: Synchrotron,
+    s4_slit_gaps: S4SlitGaps,
+    undulator: Undulator,
+    flux: Flux,
+):
+    from functools import partial
+
+    def side_set_w_return(obj, *args):
+        obj.sim_put(*args)
+        return DEFAULT
+
+    mock_omega_sets = MagicMock(
+        return_value=Status(done=True, success=True),
+        side_effect=partial(side_set_w_return, smargon.omega.user_readback),
+    )
+    mock_x = MagicMock(
+        return_value=Status(done=True, success=True),
+        side_effect=partial(side_set_w_return, smargon.x.user_readback),
+    )
+    mock_y = MagicMock(
+        return_value=Status(done=True, success=True),
+        side_effect=partial(side_set_w_return, smargon.y.user_readback),
+    )
+    mock_z = MagicMock(
+        return_value=Status(done=True, success=True),
+        side_effect=partial(side_set_w_return, smargon.z.user_readback),
+    )
+    smargon.omega.velocity.set = mock_omega_sets
+    smargon.omega.set = mock_omega_sets
+    smargon.x.set = mock_x
+    smargon.y.set = mock_y
+    smargon.z.set = mock_z
+
+    mock_arm = MagicMock(
+        side_effect=zebra.pc.arm.armed.set, return_value=Status(done=True, success=True)
+    )
+    zebra.pc.arm.arm_set.set = mock_arm
+
+    with patch("bluesky.plan_stubs.wait", autospec=True):
+        do_rotation_plan_for_tests(
+            RE,
+            mock_rotation_subscriptions,
+            undulator,
+            synchrotron,
+            s4_slit_gaps,
+            flux,
+            attenuator,
+            test_params,
+            smargon,
+            zebra,
+            backlight,
+            detector_motion,
+        )
+
+    return {
+        "RE": RE,
+        "test_rotation_params": test_params,
+        "smargon": smargon,
+        "zebra": zebra,
+        "eiger": eiger,
+        "attenuator": attenuator,
+        "detector_motion": detector_motion,
+        "backlight": backlight,
+        "mock_rotation_subscriptions": mock_rotation_subscriptions,
+        "synchrotron": synchrotron,
+        "s4_slit_gaps": s4_slit_gaps,
+        "undulator": undulator,
+        "flux": flux,
+    }
+
+
+@pytest.fixture
+def setup_and_run_rotation_plan_for_tests_standard(
+    RE: RunEngine,
+    test_rotation_params: RotationInternalParameters,
+    smargon: Smargon,
+    zebra: Zebra,
+    eiger: EigerDetector,
+    attenuator: Attenuator,
+    detector_motion: DetectorMotion,
+    backlight: Backlight,
+    mock_rotation_subscriptions: RotationCallbackCollection,
+    synchrotron: Synchrotron,
+    s4_slit_gaps: S4SlitGaps,
+    undulator: Undulator,
+    flux: Flux,
+):
+    return setup_and_run_rotation_plan_for_tests_standard(
+        RE,
+        test_rotation_params,
+        smargon,
+        zebra,
+        eiger,
+        attenuator,
+        detector_motion,
+        backlight,
+        mock_rotation_subscriptions,
+        synchrotron,
+        s4_slit_gaps,
+        undulator,
+        flux,
+    )
+
+
+@pytest.fixture
+def setup_and_run_rotation_plan_for_tests_nomove(
+    RE: RunEngine,
+    test_rotation_params_nomove: RotationInternalParameters,
+    smargon: Smargon,
+    zebra: Zebra,
+    eiger: EigerDetector,
+    attenuator: Attenuator,
+    detector_motion: DetectorMotion,
+    backlight: Backlight,
+    mock_rotation_subscriptions: RotationCallbackCollection,
+    synchrotron: Synchrotron,
+    s4_slit_gaps: S4SlitGaps,
+    undulator: Undulator,
+    flux: Flux,
+):
+    return setup_and_run_rotation_plan_for_tests_standard(
+        RE,
+        test_rotation_params_nomove,
+        smargon,
+        zebra,
+        eiger,
+        attenuator,
+        detector_motion,
+        backlight,
+        mock_rotation_subscriptions,
+        synchrotron,
+        s4_slit_gaps,
+        undulator,
+        flux,
+    )
