+++ src/artemis/experiment_plans/tests/test_rotation_scan_plan.py
@@ -1,16 +1,10 @@
 from __future__ import annotations
 
 from typing import TYPE_CHECKING
-from unittest.mock import DEFAULT, MagicMock, patch
+from unittest.mock import MagicMock, patch
 
 import pytest
 from bluesky.run_engine import RunEngine
-from bluesky.utils import Msg
-from dodal.beamlines import i03
-from dodal.devices.flux import Flux
-from dodal.devices.s4_slit_gaps import S4SlitGaps
-from dodal.devices.synchrotron import Synchrotron
-from dodal.devices.undulator import Undulator
 from ophyd.status import Status
 
 from artemis.experiment_plans.rotation_scan_plan import (
@@ -20,6 +14,7 @@ from artemis.experiment_plans.rotation_scan_plan import (
     move_to_start_w_buffer,
     rotation_scan_plan,
 )
+from artemis.experiment_plans.tests.conftest import fake_read
 from artemis.external_interaction.callbacks.rotation.rotation_callback_collection import (
     RotationCallbackCollection,
 )
@@ -57,16 +52,11 @@ def test_move_to_start(smargon: Smargon, RE):
     )
 
 
-def __fake_read(obj, initial_positions, _):
-    initial_positions[obj] = 0
-    yield Msg("null", obj)
-
-
 def test_move_to_end(smargon: Smargon, RE):
     scan_width = 153
     with patch(
         "bluesky.preprocessors.__read_and_stash_a_motor",
-        __fake_read,
+        fake_read,
     ):
         RE(
             move_to_end_w_buffer(
@@ -119,137 +109,14 @@ def test_get_plan(
     eiger.unstage.assert_called()
 
 
-def do_rotation_plan_for_tests(
-    run_engine,
-    callbacks,
-    sim_und,
-    sim_synch,
-    sim_slits,
-    sim_flux,
-    sim_att,
-    expt_params,
-    sim_sgon,
-    sim_zeb,
-    sim_bl,
-    sim_det,
-):
-    with (
-        patch(
-            "bluesky.preprocessors.__read_and_stash_a_motor",
-            __fake_read,
-        ),
-        patch(
-            "artemis.experiment_plans.rotation_scan_plan.RotationCallbackCollection.from_params",
-            lambda _: callbacks,
-        ),
-        patch("dodal.beamlines.i03.undulator", lambda: sim_und),
-        patch("dodal.beamlines.i03.synchrotron", lambda: sim_synch),
-        patch("dodal.beamlines.i03.s4_slit_gaps", lambda: sim_slits),
-        patch("dodal.beamlines.i03.flux", lambda: sim_flux),
-        patch("dodal.beamlines.i03.attenuator", lambda: sim_att),
-    ):
-        run_engine(
-            rotation_scan_plan(
-                expt_params,
-                sim_sgon,
-                sim_zeb,
-                sim_bl,
-                sim_att,
-                sim_det,
-            )
-        )
-
-
-@pytest.fixture
-def setup_and_run_rotation_plan_for_tests(
-    RE: RunEngine,
-    test_rotation_params: RotationInternalParameters,
-    smargon: Smargon,
-    zebra: Zebra,
-    eiger: EigerDetector,
-    attenuator: Attenuator,
-    detector_motion: DetectorMotion,
-    backlight: Backlight,
-    mock_rotation_subscriptions: RotationCallbackCollection,
-    synchrotron: Synchrotron,
-    s4_slit_gaps: S4SlitGaps,
-    undulator: Undulator,
-    flux: Flux,
-):
-    from functools import partial
-
-    def side_set_w_return(obj, *args):
-        obj.sim_put(*args)
-        return DEFAULT
-
-    mock_omega_sets = MagicMock(
-        return_value=Status(done=True, success=True),
-        side_effect=partial(side_set_w_return, smargon.omega.user_readback),
-    )
-    mock_x = MagicMock(
-        return_value=Status(done=True, success=True),
-        side_effect=partial(side_set_w_return, smargon.x.user_readback),
-    )
-    mock_y = MagicMock(
-        return_value=Status(done=True, success=True),
-        side_effect=partial(side_set_w_return, smargon.y.user_readback),
-    )
-    mock_z = MagicMock(
-        return_value=Status(done=True, success=True),
-        side_effect=partial(side_set_w_return, smargon.z.user_readback),
-    )
-    smargon.omega.velocity.set = mock_omega_sets
-    smargon.omega.set = mock_omega_sets
-    smargon.x.set = mock_x
-    smargon.y.set = mock_y
-    smargon.z.set = mock_z
-
-    mock_arm = MagicMock(
-        side_effect=zebra.pc.arm.armed.set, return_value=Status(done=True, success=True)
-    )
-    zebra.pc.arm.arm_set.set = mock_arm
-
-    with patch("bluesky.plan_stubs.wait", autospec=True):
-        do_rotation_plan_for_tests(
-            RE,
-            mock_rotation_subscriptions,
-            undulator,
-            synchrotron,
-            s4_slit_gaps,
-            flux,
-            attenuator,
-            test_rotation_params,
-            smargon,
-            zebra,
-            backlight,
-            detector_motion,
-        )
-
-    return {
-        "RE": RE,
-        "test_rotation_params": test_rotation_params,
-        "smargon": smargon,
-        "zebra": zebra,
-        "eiger": eiger,
-        "attenuator": attenuator,
-        "detector_motion": detector_motion,
-        "backlight": backlight,
-        "mock_rotation_subscriptions": mock_rotation_subscriptions,
-        "synchrotron": synchrotron,
-        "s4_slit_gaps": s4_slit_gaps,
-        "undulator": undulator,
-        "flux": flux,
-    }
-
-
-def test_rotation_plan_runs(setup_and_run_rotation_plan_for_tests):
-    RE: RunEngine = setup_and_run_rotation_plan_for_tests["RE"]
+def test_rotation_plan_runs(setup_and_run_rotation_plan_for_tests_standard):
+    RE: RunEngine = setup_and_run_rotation_plan_for_tests_standard["RE"]
     assert RE._exit_status == "success"
 
 
-def test_rotation_plan_zebra_settings(setup_and_run_rotation_plan_for_tests):
-    zebra: Zebra = setup_and_run_rotation_plan_for_tests["zebra"]
-    params: RotationInternalParameters = setup_and_run_rotation_plan_for_tests[
+def test_rotation_plan_zebra_settings(setup_and_run_rotation_plan_for_tests_standard):
+    zebra: Zebra = setup_and_run_rotation_plan_for_tests_standard["zebra"]
+    params: RotationInternalParameters = setup_and_run_rotation_plan_for_tests_standard[
         "test_rotation_params"
     ]
     expt_params = params.experiment_params
@@ -259,9 +126,9 @@ def test_rotation_plan_zebra_settings(setup_and_run_rotation_plan_for_tests):
     assert zebra.pc.pulse_start.get() == expt_params.shutter_opening_time_s
 
 
-def test_rotation_plan_smargon_settings(setup_and_run_rotation_plan_for_tests):
-    smargon: Smargon = setup_and_run_rotation_plan_for_tests["smargon"]
-    params: RotationInternalParameters = setup_and_run_rotation_plan_for_tests[
+def test_rotation_plan_smargon_settings(setup_and_run_rotation_plan_for_tests_standard):
+    smargon: Smargon = setup_and_run_rotation_plan_for_tests_standard["smargon"]
+    params: RotationInternalParameters = setup_and_run_rotation_plan_for_tests_standard[
         "test_rotation_params"
     ]
     expt_params = params.experiment_params
@@ -325,33 +192,6 @@ def test_cleanup_happens(
         cleanup_plan.assert_called_once()
 
 
-@pytest.fixture()
-def fake_create_devices(
-    eiger: EigerDetector,
-    smargon: Smargon,
-    zebra: Zebra,
-):
-    eiger.stage = MagicMock()
-    eiger.unstage = MagicMock()
-    mock_omega_sets = MagicMock(return_value=Status(done=True, success=True))
-
-    mock_arm_disarm = MagicMock(
-        side_effect=zebra.pc.arm.armed.set, return_value=Status(done=True, success=True)
-    )
-    zebra.pc.arm.set = mock_arm_disarm
-    smargon.omega.velocity.set = mock_omega_sets
-    smargon.omega.set = mock_omega_sets
-
-    devices = {
-        "eiger": i03.eiger(fake_with_ophyd_sim=True),
-        "smargon": smargon,
-        "zebra": zebra,
-        "detector_motion": i03.detector_motion(fake_with_ophyd_sim=True),
-        "backlight": i03.backlight(fake_with_ophyd_sim=True),
-    }
-    return devices
-
-
 @pytest.mark.s03
 @patch("bluesky.plan_stubs.wait")
 @patch("artemis.external_interaction.nexus.write_nexus.NexusWriter")
@@ -393,7 +233,7 @@ def test_ispyb_deposition_in_plan(
         patch("dodal.beamlines.i03.flux", return_value=flux),
         patch(
             "bluesky.preprocessors.__read_and_stash_a_motor",
-            __fake_read,
+            fake_read,
         ),
         patch(
             "artemis.experiment_plans.rotation_scan_plan.RotationCallbackCollection.from_params",
